Convert to Package
Add argparse
Add .env file
Add dotenv
Update gitignore
Major fix and move to package
