+++ requirements.txt
@@ -2,4 +2,6 @@ google-api-python-client
 google-auth-httplib2
 google-auth-oauthlib
 line-notify==0.1.4
-
+python-dotenv==0.21.1
+toml==0.10.2
+tomli==2.0.1
