+++ src/gmail2line/constants.py
@@ -14,12 +14,14 @@ FROM_KIDZDUO = "ie-square@1915.jp"
 FROM_GATE = "キッズセキュリティ ミマモルメ <mimamorume@hanshin-anshin.jp>"
 FROM_TRAIN = "東急セキュリティ <tskids@goopas.jp>"
 FROM_NICHINOKEN = "myn@nichinoken.co.jp"
+FROM_INSTITUTE = "東京個別指導学院 <oshirase@tkg-jp.com>"
 
 BUS_SUB = "エキッズ"
 TRAIN_SUB = "エキッズ"
 KIDZDUO_SUB = "入退室情報のお知らせ"
 GATE_SUB = "【キッズセキュリティミマモルメ】通過のお知らせ"
 NICHI_SUB = "日能研パス・サイン　メール送信"
+INSTITUTE_SUB1 = "退室のご連絡"
+INSTITUTE_SUB2 = "入室のご連絡"
 
-
-SUBJECTS = [BUS_SUB, TRAIN_SUB, KIDZDUO_SUB, GATE_SUB, NICHI_SUB]
+SUBJECTS = [BUS_SUB, TRAIN_SUB, KIDZDUO_SUB, GATE_SUB, NICHI_SUB, INSTITUTE_SUB1, INSTITUTE_SUB2]
